Fixed test naming to ensure tests are run
