+++ andhow-testing/andhow-simulated-app-tests/pom.xml
@@ -39,7 +39,6 @@
 		</pluginManagement>
 		<plugins>
 			<plugin>
-				<inherited>true</inherited>
 				<artifactId>maven-deploy-plugin</artifactId>
 				<configuration>
 					<skip>true</skip>
